update tests that created int64 images

+++ nilearn/plotting/tests/test_find_cuts.py
@@ -52,7 +52,7 @@ def test_find_cut_coords():
     data = np.zeros((20, 20, 20))
     data[4:6, 4:6, 4:6] = 1000
     img = nibabel.Nifti1Image(data, 2 * np.eye(4))
-    mask_data = np.ones((20, 20, 20), dtype=int)
+    mask_data = np.ones((20, 20, 20), dtype="uint8")
     mask_img = nibabel.Nifti1Image(mask_data, 2 * np.eye(4))
     cut_coords = find_xyz_cut_coords(img, mask_img=mask_img)
     np.testing.assert_array_equal(cut_coords,
@@ -72,7 +72,7 @@ def test_find_cut_coords():
 
     # Check that a warning is given when all values are masked
     # due to thresholding and that the center of mass is returned
-    mask_data = np.ones((20, 20, 20), dtype=int)
+    mask_data = np.ones((20, 20, 20), dtype="uint8")
     mask_img = nibabel.Nifti1Image(mask_data, np.eye(4))
     with pytest.warns(UserWarning,
                       match=("Could not determine cut coords: "
